Updated `lib/netzke-core` to require the `hobo_support` gem early so Netzke can be used in a Rails application that depends on Hobo <http://hobocentral.net>
Hobo includes a bunch of extensions and patches to various Rails and Ruby classes including ActiveRecord, Array and Hash in it's `hobo_support` gem.

To let both libraries co-exist nicely let's make sure that the Hobo extensions and patches to Rails and Ruby classes get loaded before requiring Netzke's extensions and patches.

This update should have not influence on plain Rails applications because the `require` statement is guarded by `if defined? HoboSupport`.