Also check for icons in the assets.
Netzke now uses `assets/icons` as the default path
to find the icons, so we need to look for the icons
in the asset-folders. Since the icons are supplied
by a vendor, I place those generally in `vendor/assets`
so for completeness check all asset paths.

This will still fail if the icons are bundled in a
asset gem.